be: renamed and refined user-exists
